requirements.txt
alembic==1.7.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
contextlib2==21.6.0
Deprecated==1.2.13
Flask==2.1.1
Flask-Migrate==3.1.0
Flask-OpenTracing==1.1.0
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
itsdangerous==2.1.2
jaeger-client==4.8.0
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
opentelemetry-api==1.10.0
opentelemetry-instrumentation==0.30b1
opentelemetry-instrumentation-dbapi==0.29b0
opentelemetry-instrumentation-flask==0.30b1
opentelemetry-instrumentation-psycopg2==0.29b0
opentelemetry-instrumentation-wsgi==0.30b1
opentelemetry-semantic-conventions==0.30b1
opentelemetry-util-http==0.30b1
opentracing==2.4.0
opentracing-instrumentation==3.3.1
psycopg2-binary==2.9.3
python-dotenv==0.20.0
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.35
threadloop==1.0.2
thrift==0.16.0
tornado==5.1.1
urllib3==1.26.9
Werkzeug==2.1.1
wrapt==1.14.0
zipp==3.8.0
